refactor(github): replace MCP client with direct Git operations
- Remove GitHub MCP client implementation and tests
- Create simplified GitHub client using direct Git operations
- Move test code to dedicated tests directory
- Update development guide to reflect the non-MCP approach

This change improves reliability by using standard Git commands and GitHub API
directly instead of depending on the MCP server integration, which can be
implemented as a future enhancement if needed.
